data loading passing tests
